Updated README.md and fixed failing tests because of moved header file.
