+++ README.md
@@ -1,16 +1,23 @@
 # packager
 
-QNX C program for encoding sensor data into radio packets.
+C program for encoding sensor data into CU InSpace radio packet format.
 
 ## Usage
 
-This software is distributed under the [GNUGPLv3 license](https://github.com/CarletonURocketry/packager/blob/main/LICENSE), which requires all source code in this repository
-to be made publicly available if this software is distributed. Additionally, this license requires prominent attribution
-of the original author(s). It's encouraged to read the entire license for more information.
+`packager` expects input data coming from a message queue called `/fetcher/sensors`. See [fetcher][fetcher] for more 
+details on the format of these messages.
+
+For more information about how to use packager from the command line, type `use ./packager` in your shell.
 
 ## Installation
 
 - Requires QNX 7.1
-  - `qcc`
-  - QNX C libraries included in the 7.1 distribution
-  - GNUMake
+
+## Notes
+
+This software is distributed under the [GNUGPLv3 license](https://github.com/CarletonURocketry/packager/blob/main/LICENSE),
+which requires all source code in this repository to be made publicly available if this software is distributed.
+Additionally, this license requires prominent attribution of the original author(s). It's encouraged to read the entire
+license for more information.
+
+[fetcher]: https://github.com/CarletonURocketry/fetcher
